requirements.txt
bcrypt
pydantic
PyJWT
python-rapidjson
sanic
tortoise-orm
requests